revise test of component in page_index for new folder structure
