+++ client-server/test/components/page_index/index.spec.js
@@ -1,10 +1,9 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 import { expect } from 'chai';
-import Index from '../../app/components/index.jsx';
-import Header from '../../app/components/header.jsx';
-import ElementList from '../../app/components/elementList.jsx';
-import EditModal from '../../app/components/editModal.jsx';
+import Index from 'components/page_index/index.jsx';
+import Header from 'components/page_index/header.jsx';
+import ElementList from 'components/page_index/elementList.jsx';
 
 
 describe('<Index />', () => {
@@ -19,10 +18,6 @@ describe('<Index />', () => {
 
     it('should render one <ElementList /> component', () => {
       expect(wrapper.find(ElementList)).to.have.length(1);
-    });
-
-    it('should render one <EditModal> component', ()=> {
-      expect(wrapper.find(EditModal)).to.have.length(1);
     });
   });
 });
